requirements.txt
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
DateTime==5.5
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
lxml==5.2.1
pydantic==2.7.0
pydantic_core==2.18.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.1.0
zope.interface==6.3
